moved concepts and traits - moved converter concepts to converters project - moved clipboard concept and traits to clipboard project - updated cli to include new headers from other projects
