+++ base64.cli/base64.cli.vcxproj
@@ -110,9 +110,11 @@
   <ItemGroup>
     <ClCompile Include="application.cpp" />
     <ClInclude Include="base64_app.h" />
-    <ClCompile Include="base64_app.cpp" >
+    <ClCompile Include="base64_app.cpp">
       <DependentUpon>base64_app.h</DependentUpon>
     </ClCompile>
+    <ClInclude Include="operation_type.h" />
+    <ClInclude Include="output_type.h" />
     <ClInclude Include="pch.h" />
     <ClCompile Include="pch.cpp">
       <DependentUpon>pch.h</DependentUpon>
