+++ base64.cli/base64_app.h
@@ -16,72 +16,19 @@
 #include <span>
 #include <string>
 #include <string_view>
-#include <optional>
 #include <vector>
 
+#include "output_type.h"
+#include "operation_type.h"
+
+#include "../base64.converters/concepts.h"
 #include "../base64.converters/decoder.h"
 #include "../base64.converters/encoder.h"
-#include "../base64.converters/maybe_encoded.h"
-#include "../modern_win32_api.user/clipboard.h"
+#include "../modern_win32_api.user/clipboard_concept.h"
+#include "../modern_win32_api.user/clipboard_traits.h"
 
 namespace moreland::base64::cli
 {
-    using byte_string = std::basic_string<unsigned char>;
-
-    using converters::encoder;
-    using converters::decoder;
-
-    template <typename TEncoder>
-    concept Encoder = requires(TEncoder const& encoder, std::span<unsigned char const> const source, int x)
-    {
-        { encoder.encode(source) } -> std::convertible_to<std::optional<std::vector<unsigned char >>>;
-        { encoder.encode_to_string_or_empty(source) } -> std::convertible_to<std::string>;
-    };
-
-    template <typename TDecoder>
-    concept Decoder = requires(TDecoder const& decoder, std::span<unsigned char const> const source)
-    {
-        { decoder.decode(source) } -> std::convertible_to<converters::maybe_encoded<std::vector<unsigned char >>>;
-        { decoder.decode_to_string_or_empty(source) } -> std::convertible_to<std::string>;
-    };
-
-    template <typename TClipboard>
-    concept Clipboard = requires(std::string_view const data)
-    {
-        { TClipboard::get_clipboard() } -> std::convertible_to<std::optional<std::string>>;
-        { TClipboard::set_clipboard(data) } -> std::convertible_to<bool>;
-    };
-
-    enum class operation_type
-    {
-        unknown,
-        decode,
-        encode,
-    };
-
-    enum class output_type
-    {
-        unknown,
-        clipboard,
-        file_to_clipboard,
-        file_to_file,
-    };
-
-    struct clipboard_traits
-    {
-        [[nodiscard]]
-        static std::optional<std::string> get_clipboard() 
-        {
-            return modern_win32_api::user::get_clipboard();
-        }
-
-        [[nodiscard]]
-        static bool set_clipboard(std::string_view const data)
-        {
-            return modern_win32_api::user::set_clipboard(data);
-        }
-    };
-
 
     [[nodiscard]]
     bool string_lower_equals(std::string_view const first, std::string_view const second);
@@ -92,10 +39,15 @@ namespace moreland::base64::cli
     [[nodiscard]]
     std::vector<std::string_view> as_vector_of_views(char const* source[], std::size_t const length);
 
-    template <Encoder ENCODER = encoder, Decoder DECODER = decoder, Clipboard CLIPBOARD = clipboard_traits>
+    template <
+        converters::Encoder ENCODER = converters::encoder,
+        converters::Decoder DECODER = converters::decoder,
+        modern_win32_api::user::Clipboard CLIPBOARD = modern_win32_api::user::clipboard_traits
+    >
     [[nodiscard]]
     bool base64_run(std::span<std::string_view const> const arguments, ENCODER const& encoder, DECODER const& decoder)
     {
+        using byte_string = std::basic_string<unsigned char>;
         auto [operation, output] = get_operation_and_output_from_arguments(arguments);
 
         if (operation == operation_type::unknown) {
