+++ base64.converters/base64.converters.vcxproj
@@ -92,6 +92,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="base64_failure_reason.h" />
+    <ClInclude Include="concepts.h" />
     <ClInclude Include="decoder.h" />
     <ClCompile Include="decoder.cpp">
       <DependentUpon>decoder.h</DependentUpon>
