+++ base64.converters/base64.converters.vcxproj.filters
@@ -13,5 +13,6 @@
     <ClInclude Include="decoder.h" />
     <ClInclude Include="base64_failure_reason.h" />
     <ClInclude Include="maybe_encoded.h" />
+    <ClInclude Include="concepts.h" />
   </ItemGroup>
 </Project>
